requirements.txt
click>=8.0.0
requests>=2.28.0
pyyaml>=6.0
python-dotenv>=0.19.0
ffmpeg-python>=0.2.0
tmdbv3api>=1.7.0
